intelligence/requirements.txt
flask
transformers
torch
spacy
python-dotenv
requests
dateparser
